requirements.txt
streamlit==1.24.0
google-cloud-storage==2.14.0
langchain==0.1.14
chromadb==0.4.24
google-cloud-aiplatform==1.44.0
langchain-google-vertexai==0.1.2